refactor: split reviews and social components
